% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCohortCharacteristics.R
\name{getCohortCharacteristics}
\alias{getCohortCharacteristics}
\title{Create characterization of a cohort}
\usage{
getCohortCharacteristics(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  oracleTempSchema = NULL,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort",
  cohortId,
  covariateSettings,
  minCellCount
)
}
\arguments{
\item{cohortId}{The cohort definition ID used to reference the cohort in the cohort
table.}

\item{covariateSettings}{Either an object of type \code{covariateSettings} as created using one of
the createCovariate functions in the FeatureExtraction package, or a list
of such objects.}

\item{minCellCount}{The minimum cell count that should be returned}

\item{Connection}{connection}

\item{CdmDatabaseSchema}{cdmDatabaseSchema}

\item{OracleTempSchema}{OracleTempSchema}

\item{CohortTable}{CohortTable}
}
\value{
A list of data frames with cohort characteristics.
}
\description{
Computes features using all drugs, conditions, procedures, etc. observed on or prior to the cohort
index date.
}
